refactor: remove unused code in detailactivityview and refactor mapview to group component
